pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phase-engine"
version = "1.3.0"
description = "Coherent HF Phased Array Middleware for GPSDO-locked SDRs"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Michael James Hauan", email = "ac0g@arrl.net"}
]
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Communications :: Ham Radio",
]
keywords = ["sdr", "phased-array", "beamforming", "hf", "radio", "dsp"]

dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "toml>=0.10",
    "ka9q-python @ git+https://github.com/mijahauan/ka9q-python.git"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "mypy>=1.0",
]
plot = [
    "matplotlib>=3.5.0",
]

[project.scripts]
phase-engine = "phase_engine.cli:main"

[project.urls]
Homepage = "https://github.com/mijahauan/phase-engine"
Repository = "https://github.com/mijahauan/phase-engine"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
